Moved context logic to auth feature. Created Provider and hook to use it
