+++ src/features/auth/context/useAuthContext.ts
@@ -1,0 +1,8 @@
+import { useContext } from "react";
+import { AuthContext } from "./AuthContext";
+
+export const useAuthContext = () => {
+  const { user, setUser } = useContext(AuthContext);
+
+  return { user, setUser };
+};
